check ticket class is used

+++ DEMO/BLL/Ticket_Class_BLL.cs
@@ -6,11 +6,13 @@ using System.Linq;
 using System.Security.Policy;
 using System.Text;
 using System.Threading.Tasks;
+using System.Web;
 
 namespace BLL
 {
     public class Ticket_Class_BLL
     {
+        string state = string.Empty;
         public List<TicketClassDTO> L_TicketClass()
         {
             /*Ticket_classAccess ticket_class = new Ticket_classAccess();
@@ -23,11 +25,18 @@ namespace BLL
         }
         public int DeleteTicketClass(string ID)
         {
+            if(new DAL.Ticket_classAccess().isUsed(ID))
+            {
+                state = "Hạng vé này đang được sử dụng";
+                return 0;
+            }
             return new DAL.Ticket_classAccess().DeleteTicketClass(ID);
         }
         public string GetState()
         {
-            return new DAL.Ticket_classAccess().GetState();
+            if(new DAL.Ticket_classAccess().GetState() != string.Empty)
+                return new DAL.Ticket_classAccess().GetState();
+            return state;
         }
     }
 }
